feat: 이미지 인식 완료
- 이미지 파일 업로드 후 단어 인식
