--- src/main/java/com/example/marimo_back/TestController.java
@@ -1,158 +1,0 @@
-package com.example.marimo_back;
-
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.cloud.storage.*;
-import com.google.cloud.translate.*;
-import com.google.cloud.vision.v1.*;
-import com.google.cloud.vision.v1.Image;
-import com.google.common.collect.Lists;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.web.bind.annotation.*;
-
-import com.google.api.gax.paging.Page;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.awt.*;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-import java.io.IOException;
-
-
-@RestController
-@RequiredArgsConstructor
-@CrossOrigin(origins = "http://localhost:3000")
-public class TestController {
-
-    @GetMapping(value = "/test")
-    /**
-     * Detects localized objects in a remote image on Google Cloud Storage.
-     *
-     * @param gcsPath The path to the remote file on Google Cloud Storage to detect localized objects
-     *     on.
-     * @throws Exception on errors while closing the client.
-     * @throws IOException on Input/Output errors.
-     */
-    public static void detectLocalizedObjectsGcs() throws IOException {
-
-//        authExplicit("src/main/resources/static/top-glass-322515-5fcccf54600d.json");
-        String gcsPath="gs://marimo_bucket/test.png";
-        List<AnnotateImageRequest> requests = new ArrayList<>();
-
-        ImageSource imgSource = ImageSource.newBuilder().setGcsImageUri(gcsPath).build();
-        Image img = Image.newBuilder().setSource(imgSource).build();
-
-        AnnotateImageRequest request =
-                AnnotateImageRequest.newBuilder()
-                        .addFeatures(Feature.newBuilder().setType(Feature.Type.OBJECT_LOCALIZATION))
-                        .setImage(img)
-                        .build();
-        requests.add(request);
-
-        // Initialize client that will be used to send requests. This client only needs to be created
-        // once, and can be reused for multiple requests. After completing all of your requests, call
-        // the "close" method on the client to safely clean up any remaining background resources.
-        try (ImageAnnotatorClient client = ImageAnnotatorClient.create()) {
-            // Perform the request
-            BatchAnnotateImagesResponse response = client.batchAnnotateImages(requests);
-            List<AnnotateImageResponse> responses = response.getResponsesList();
-            client.close();
-            // Display the results
-            for (AnnotateImageResponse res : responses) {
-                for (LocalizedObjectAnnotation entity : res.getLocalizedObjectAnnotationsList()) {
-                    System.out.format("Object name: %s%n", entity.getName());
-                    System.out.format("Confidence: %s%n", entity.getScore());
-                    System.out.format("Normalized Vertices:%n");
-                    entity
-                            .getBoundingPoly()
-                            .getNormalizedVerticesList()
-                            .forEach(vertex -> System.out.format("- (%s, %s)%n", vertex.getX(), vertex.getY()));
-
-//                    System.out.println(entity.getName());
-                    Translate translate = TranslateOptions.getDefaultInstance().getService();
-                    Translation translation = translate.translate(entity.getName(), Translate.TranslateOption.targetLanguage("ko"));
-                    System.out.println(translation.getTranslatedText());
-                }
-            }
-        }
-    }
-
-    @PostMapping("/images")
-    public String uploadImage(@RequestPart(value = "images", required = false) MultipartFile file) throws IOException {
-        System.out.println("이미지업로드");
-        // The ID of your GCP project
-         String projectId = "top-glass-322515";
-
-        // The ID of your GCS bucket
-         String bucketName = "marimo_bucket";
-
-        // The ID of your GCS object
-        String fileName = file.getOriginalFilename();
-        SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyyMMdd"); // 폴더 이름 업로드 날짜로 바꾸기
-        String objectName = simpleDateFormat.format(new Date()) + Long.toString(System.nanoTime()); // 파일명 변경
-
-        // The path to your file to upload
-        try (
-                FileOutputStream fos = new FileOutputStream("/tmp/" +file.getOriginalFilename());
-                // 파일 저장할 경로 + 파일명을 파라미터로 넣고 fileOutputStream 객체 생성하고
-                InputStream is = file.getInputStream();) {
-            // file로 부터 inputStream을 가져온다.
-
-            int readCount = 0;
-            byte[] buffer = new byte[1024];
-            // 파일을 읽을 크기 만큼의 buffer를 생성하고
-            // ( 보통 1024, 2048, 4096, 8192 와 같이 배수 형식으로 버퍼의 크기를 잡는 것이 일반적이다.)
-
-            while ((readCount = is.read(buffer)) != -1) {
-                //  파일에서 가져온 fileInputStream을 설정한 크기 (1024byte) 만큼 읽고
-                fos.write(buffer, 0, readCount);
-                // 위에서 생성한 fileOutputStream 객체에 출력하기를 반복한다
-            }
-        } catch (Exception ex) {
-            throw new RuntimeException("file Save Error");
-        }
-
-        String filePath = "/tmp/" +file.getOriginalFilename();
-
-
-        Storage storage = StorageOptions.newBuilder().setProjectId(projectId).build().getService();
-        BlobId blobId = BlobId.of(bucketName, objectName);
-        BlobInfo blobInfo = BlobInfo.newBuilder(blobId).build();
-        storage.create(blobInfo, Files.readAllBytes(Paths.get(filePath)));
-
-        System.out.println(
-                "File " + filePath + " uploaded to bucket " + bucketName + " as " + objectName);
-        return "upload Image";
-    }
-
-    @GetMapping("/test2")
-    public String test() {
-        System.out.println("성공");
-        return "test";
-    }
-
-
-
-
-/*    @ConfigurationProperties("spring.cloud.gcp.vision")
-    static void authExplicit(String jsonPath) throws IOException {
-        // You can specify a credential file by providing a path to GoogleCredentials.
-        // Otherwise credentials are read from the GOOGLE_APPLICATION_CREDENTIALS environment variable.
-        GoogleCredentials credentials = GoogleCredentials.fromStream(new FileInputStream(jsonPath))
-                .createScoped(Lists.newArrayList("https://www.googleapis.com/auth/cloud-platform"));
-        Storage storage = StorageOptions.newBuilder().setCredentials(credentials).build().getService();
-
-        System.out.println("Buckets:");
-        Page<Bucket> buckets = storage.list();
-        for (Bucket bucket : buckets.iterateAll()) {
-            System.out.println(bucket.toString());
-        }
-    }*/
-}
